Ei crashi kui proovid olemasolevat ümbernimetada samaks

+++ TelesarjadeRenameimine/TaasNimetamine.cs
@@ -46,11 +46,25 @@ namespace TelesarjadeRenameimine
                 //Siin on if, sest kui sisestad S01E*, siis see näeb välja S01E1 kuni S01E10, kuid nii on see S01E01 kuni S01E10 (palju ilusam)
                 if (EsimeneOsa < 10)
                 {
-                    File.Move(file, path + nimi + "0" + EsimeneOsa + failiTüüp);
+                    if (File.Exists(path + nimi + "0" + EsimeneOsa + failiTüüp))
+                    {
+                        Console.WriteLine(path + nimi + "0" + EsimeneOsa + failiTüüp + " juba eksisteerib, jääb vahele");
+                    }
+                    else
+                    {
+                        File.Move(file, path + nimi + "0" + EsimeneOsa + failiTüüp);
+                    }
                 }
                 else
                 {
-                    File.Move(file, path + nimi + EsimeneOsa + failiTüüp);
+                    if (File.Exists(path + nimi + EsimeneOsa + failiTüüp))
+                    {
+                        Console.WriteLine(path + nimi + EsimeneOsa + failiTüüp + " juba eksisteerib, jääb vahele");
+                    }
+                    else
+                    {
+                        File.Move(file, path + nimi + EsimeneOsa + failiTüüp);
+                    }
                 }
                 EsimeneOsa++;
             }
